Fix symlink issue
We mounted a folder under /etc as volume into the nginx container to use certs. But those certs were symlinks (to accommodate for letsencrypt's updating mechanism). These symlinks don't work in do>
So we made copies of the originals into the opt folder. Means we need to regularly renew certs.
renew-certs.sh should do it easily.